+++ tests/trycmd-tests.rs
@@ -1,0 +1,4 @@
+#[test]
+fn fs_tests() {
+    trycmd::TestCases::new().case("tests/trycmd/fs/*.toml");
+}