+++ CHANGELOG.md
@@ -22,6 +22,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Validate object spec versions are monotonically increasing
 - Validate only one object/root version declaration
 - Validate all manifest entries are used at least once
+- `info` command for display object/repo details
 
 ### Fixed